requirements.txt
mysql-connector==2.2.9
tabulate==0.8.9
